advanced/cat vs dog/requirements.txt
tensorflow>=2.10.0
streamlit>=1.28.0
pillow>=9.0.0
numpy>=1.21.0
requests>=2.25.0
streamlit-lottie>=0.0.3
